update e2e test
