+++ tests/e2e_tests/subcommands/delegation/test_set_delegate_take.py
@@ -8,7 +8,7 @@ from tests.e2e_tests.utils import setup_wallet
 
 def test_set_delegate_increase_take(local_chain):
     # Register root as Alice
-    keypair, exec_command, wallet_path = setup_wallet("//Alice")
+    keypair, exec_command, wallet = setup_wallet("//Alice")
     exec_command(RootRegisterCommand, ["root", "register"])
 
     # Create subnet 1 and verify created successfully
@@ -20,7 +20,7 @@ def test_set_delegate_increase_take(local_chain):
     assert local_chain.query("SubtensorModule", "NetworksAdded", [1]).serialize()
 
     # Register and nominate Bob
-    keypair, exec_command, wallet_path = setup_wallet("//Bob")
+    keypair, exec_command, wallet = setup_wallet("//Bob")
     assert (
         local_chain.query(
             "SubtensorModule", "LastTxBlock", [keypair.ss58_address]
